DODGY: Until a file is downloaded, say it's 1TB
Tools like zcat didn't like being told that a file would be 0 bytes
and then it turn out to be more.  Less doesn't seem to like being
told that a file is 1TB and then turn out to be less, but it at
least shows you some file.  I don't know if this is going to
horribly break for a tool that takes st_size at face value